Corrected the titles in the navbar
