+++ front-end-side/src/Components/Navbar.js
@@ -10,9 +10,8 @@ const Navbar = () => {
                     <div id="links">
                         <Link to="/">Home</Link>
                         <Link to="/friends">Friends</Link>
-                        <Link to="/enter-data">Enter data</Link>
-                        <Link to="/read-from-file-data">Data from file</Link>
-                        <Link to="/show-data" class="right-link">Show data</Link>
+                        <Link to="/data">Data</Link>
+                        <Link to="/enter-data" class="right-link">Enter data</Link>
                         <Link to="/file-upload">File upload</Link>
                     </div>
                 </div>
