+++ front-end-side/src/Pages/EnterData.css
